+++ pac4j-core/src/main/java/org/pac4j/core/authorization/checker/DefaultAuthorizationChecker.java
@@ -1,8 +1,6 @@
 package org.pac4j.core.authorization.checker;
 
 import org.pac4j.core.authorization.authorizer.*;
-import org.pac4j.core.authorization.authorizer.csrf.*;
-import org.pac4j.core.context.DefaultAuthorizers;
 import org.pac4j.core.context.Pac4jConstants;
 import org.pac4j.core.context.WebContext;
 import org.pac4j.core.profile.UserProfile;
@@ -12,8 +10,6 @@ import org.slf4j.LoggerFactory;
 import java.util.*;
 
 import static org.pac4j.core.util.CommonHelper.*;
-
-import static org.pac4j.core.context.HttpConstants.*;
 
 /**
  * Default way to check the authorizations (with default authorizers).
@@ -25,95 +21,49 @@ public class DefaultAuthorizationChecker implements AuthorizationChecker {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(DefaultAuthorizationChecker.class);
 
-    final static StrictTransportSecurityHeader STRICT_TRANSPORT_SECURITY_HEADER = new StrictTransportSecurityHeader();
-    final static XContentTypeOptionsHeader X_CONTENT_TYPE_OPTIONS_HEADER = new XContentTypeOptionsHeader();
-    final static XFrameOptionsHeader X_FRAME_OPTIONS_HEADER = new XFrameOptionsHeader();
-    final static XSSProtectionHeader XSS_PROTECTION_HEADER = new XSSProtectionHeader();
-    final static CacheControlHeader CACHE_CONTROL_HEADER = new CacheControlHeader();
-    final static CsrfAuthorizer CSRF_AUTHORIZER = new CsrfAuthorizer();
-    final static CsrfTokenGeneratorAuthorizer CSRF_TOKEN_GENERATOR_AUTHORIZER
-        = new CsrfTokenGeneratorAuthorizer(new DefaultCsrfTokenGenerator());
-    final static CorsAuthorizer CORS_AUTHORIZER = new CorsAuthorizer();
-    final static IsAnonymousAuthorizer IS_ANONYMOUS_AUTHORIZER = new IsAnonymousAuthorizer();
-    final static IsAuthenticatedAuthorizer IS_AUTHENTICATED_AUTHORIZER =new IsAuthenticatedAuthorizer();
-    final static IsFullyAuthenticatedAuthorizer IS_FULLY_AUTHENTICATED_AUTHORIZER = new IsFullyAuthenticatedAuthorizer();
-    final static IsRememberedAuthorizer IS_REMEMBERED_AUTHORIZER = new IsRememberedAuthorizer();
-
-    static {
-        CORS_AUTHORIZER.setAllowOrigin("*");
-        CORS_AUTHORIZER.setAllowCredentials(true);
-        final Set<HTTP_METHOD> methods = new HashSet<>();
-        methods.add(HTTP_METHOD.GET);
-        methods.add(HTTP_METHOD.PUT);
-        methods.add(HTTP_METHOD.POST);
-        methods.add(HTTP_METHOD.DELETE);
-        methods.add(HTTP_METHOD.OPTIONS);
-        CORS_AUTHORIZER.setAllowMethods(methods);
-    }
+    static final CsrfAuthorizer CSRF_AUTHORIZER = new CsrfAuthorizer();
+    static final IsAnonymousAuthorizer IS_ANONYMOUS_AUTHORIZER = new IsAnonymousAuthorizer();
+    static final IsAuthenticatedAuthorizer IS_AUTHENTICATED_AUTHORIZER =new IsAuthenticatedAuthorizer();
+    static final IsFullyAuthenticatedAuthorizer IS_FULLY_AUTHENTICATED_AUTHORIZER = new IsFullyAuthenticatedAuthorizer();
+    static final IsRememberedAuthorizer IS_REMEMBERED_AUTHORIZER = new IsRememberedAuthorizer();
 
     @Override
     public boolean isAuthorized(final WebContext context, final List<UserProfile> profiles, final String authorizersValue,
                                 final Map<String, Authorizer> authorizersMap) {
         final List<Authorizer> authorizers = new ArrayList<>();
         String authorizerNames = authorizersValue;
-        // the authorizers are not defined, we default to:
-        if (authorizerNames == null) {
-            authorizerNames = DefaultAuthorizers.CSRF + Pac4jConstants.ELEMENT_SEPARATOR + DefaultAuthorizers.SECURITYHEADERS;
+        // no authorizers defined, we default to CSRF_CHECK
+        if (isBlank(authorizerNames)) {
+            authorizerNames = DefaultAuthorizers.CSRF_CHECK;
         }
-
-        // if we have an authorizer name (which may be a list of authorizer names)
-        if (isNotBlank(authorizerNames)) {
-            final String[] names = authorizerNames.split(Pac4jConstants.ELEMENT_SEPARATOR);
-            final int nb = names.length;
-            for (int i = 0; i < nb; i++) {
-                final String name = names[i].trim();
-                if (DefaultAuthorizers.HSTS.equalsIgnoreCase(name)) {
-                    authorizers.add(STRICT_TRANSPORT_SECURITY_HEADER);
-                } else if (DefaultAuthorizers.NOSNIFF.equalsIgnoreCase(name)) {
-                    authorizers.add(X_CONTENT_TYPE_OPTIONS_HEADER);
-                } else if (DefaultAuthorizers.NOFRAME.equalsIgnoreCase(name)) {
-                    authorizers.add(X_FRAME_OPTIONS_HEADER);
-                } else if (DefaultAuthorizers.XSSPROTECTION.equalsIgnoreCase(name)) {
-                    authorizers.add(XSS_PROTECTION_HEADER);
-                } else if (DefaultAuthorizers.NOCACHE.equalsIgnoreCase(name)) {
-                    authorizers.add(CACHE_CONTROL_HEADER);
-                } else if (DefaultAuthorizers.SECURITYHEADERS.equalsIgnoreCase(name)) {
-                    authorizers.add(CACHE_CONTROL_HEADER);
-                    authorizers.add(X_CONTENT_TYPE_OPTIONS_HEADER);
-                    authorizers.add(STRICT_TRANSPORT_SECURITY_HEADER);
-                    authorizers.add(X_FRAME_OPTIONS_HEADER);
-                    authorizers.add(XSS_PROTECTION_HEADER);
-                } else if (DefaultAuthorizers.CSRF_TOKEN.equalsIgnoreCase(name)) {
-                    authorizers.add(CSRF_TOKEN_GENERATOR_AUTHORIZER);
-                } else if (DefaultAuthorizers.CSRF_CHECK.equalsIgnoreCase(name)) {
-                    authorizers.add(CSRF_AUTHORIZER);
-                } else if (DefaultAuthorizers.CSRF.equalsIgnoreCase(name)) {
-                    authorizers.add(CSRF_TOKEN_GENERATOR_AUTHORIZER);
-                    authorizers.add(CSRF_AUTHORIZER);
-                } else if (DefaultAuthorizers.ALLOW_AJAX_REQUESTS.equalsIgnoreCase(name)) {
-                    authorizers.add(CORS_AUTHORIZER);
-                } else if (DefaultAuthorizers.IS_ANONYMOUS.equalsIgnoreCase(name)) {
-                    authorizers.add(IS_ANONYMOUS_AUTHORIZER);
-                } else if (DefaultAuthorizers.IS_AUTHENTICATED.equalsIgnoreCase(name)) {
-                    authorizers.add(IS_AUTHENTICATED_AUTHORIZER);
-                } else if (DefaultAuthorizers.IS_FULLY_AUTHENTICATED.equalsIgnoreCase(name)) {
-                    authorizers.add(IS_FULLY_AUTHENTICATED_AUTHORIZER);
-                } else if (DefaultAuthorizers.IS_REMEMBERED.equalsIgnoreCase(name)) {
-                    authorizers.add(IS_REMEMBERED_AUTHORIZER);
-                } else {
-                    // we must have authorizers
-                    assertNotNull("authorizersMap", authorizersMap);
-                    Authorizer result = null;
-                    for (final Map.Entry<String, Authorizer> entry : authorizersMap.entrySet()) {
-                        if (areEqualsIgnoreCaseAndTrim(entry.getKey(), name)) {
-                            result = entry.getValue();
-                            break;
-                        }
+        final String[] names = authorizerNames.split(Pac4jConstants.ELEMENT_SEPARATOR);
+        final int nb = names.length;
+        for (int i = 0; i < nb; i++) {
+            final String name = names[i].trim();
+            if (DefaultAuthorizers.CSRF_CHECK.equalsIgnoreCase(name)) {
+                authorizers.add(CSRF_AUTHORIZER);
+            } else if (DefaultAuthorizers.IS_ANONYMOUS.equalsIgnoreCase(name)) {
+                authorizers.add(IS_ANONYMOUS_AUTHORIZER);
+            } else if (DefaultAuthorizers.IS_AUTHENTICATED.equalsIgnoreCase(name)) {
+                authorizers.add(IS_AUTHENTICATED_AUTHORIZER);
+            } else if (DefaultAuthorizers.IS_FULLY_AUTHENTICATED.equalsIgnoreCase(name)) {
+                authorizers.add(IS_FULLY_AUTHENTICATED_AUTHORIZER);
+            } else if (DefaultAuthorizers.IS_REMEMBERED.equalsIgnoreCase(name)) {
+                authorizers.add(IS_REMEMBERED_AUTHORIZER);
+            // we don't add any authorizer for none
+            } else if (!DefaultAuthorizers.NONE.equalsIgnoreCase(name)){
+                // we must have authorizers
+                assertNotNull("authorizersMap", authorizersMap);
+                Authorizer result = null;
+                for (final Map.Entry<String, Authorizer> entry : authorizersMap.entrySet()) {
+                    if (areEqualsIgnoreCaseAndTrim(entry.getKey(), name)) {
+                        result = entry.getValue();
+                        break;
                     }
-                    // we must have an authorizer defined for this name
-                    assertNotNull("authorizersMap['" + name + "']", result);
-                    authorizers.add(result);
                 }
+                // we must have an authorizer defined for this name
+                assertNotNull("authorizersMap['" + name + "']", result);
+                authorizers.add(result);
             }
         }
         return isAuthorized(context, profiles, authorizers);
