+++ pac4j-core/src/main/java/org/pac4j/core/matching/matcher/HttpMethodMatcher.java
@@ -1,4 +1,4 @@
-package org.pac4j.core.matching;
+package org.pac4j.core.matching.matcher;
 
 import org.pac4j.core.context.WebContext;
 import org.pac4j.core.util.CommonHelper;
