+++ pac4j-core/src/main/java/org/pac4j/core/matching/matcher/csrf/CsrfTokenGenerator.java
@@ -1,4 +1,4 @@
-package org.pac4j.core.authorization.authorizer.csrf;
+package org.pac4j.core.matching.matcher.csrf;
 
 import org.pac4j.core.context.WebContext;
